fixed: Ticket #4822 - Refresh TV episode from xml nfo will create multiple DB IDs
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@15540 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/VideoDatabase.cpp
@@ -2317,6 +2317,10 @@ void CVideoDatabase::DeleteEpisode(const CStdString& strFilenameAndPath, long lE
       }
     }
 
+    long lFileId = GetFileId(strFilenameAndPath);
+    if (lFileId < 0)
+      return ;
+
     CStdString strSQL;
     strSQL=FormatSQL("delete from actorlinkepisode where idepisode=%i", lEpisodeId);
     m_pDS->exec(strSQL.c_str());
@@ -2339,7 +2343,7 @@ void CVideoDatabase::DeleteEpisode(const CStdString& strFilenameAndPath, long lE
     {
       ClearBookMarksOfFile(strFilenameAndPath);
 
-      strSQL=FormatSQL("delete from episode where idepisode=%i", lEpisodeId);
+      strSQL=FormatSQL("delete from episode where idfile=%i", lFileId);
       m_pDS->exec(strSQL.c_str());
     }
     /*
